RP-366 Create Sandbox API/Service Moving the AggregateReporting configuration files to rdw_reporting_common first commit with successful IT build.
