+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/AggregateReportingConfiguration.java
@@ -1,8 +1,8 @@
 package org.opentestsystem.rdw.olap;
 
 import org.opentestsystem.rdw.multitenant.TenantKeyResolver;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingPropertiesResolver;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingPropertiesRoot;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesResolver;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesRoot;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.cloud.context.config.annotation.RefreshScope;
 import org.springframework.context.annotation.Bean;
