+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/ApplicationConfiguration.java
@@ -24,7 +24,7 @@ import org.opentestsystem.rdw.multitenant.TenantProperties;
 import org.opentestsystem.rdw.multitenant.datasource.DataSourceElementsProperties;
 import org.opentestsystem.rdw.multitenant.messaging.TenantChannelInterceptor;
 import org.opentestsystem.rdw.multitenant.messaging.TenantMessageRequestResolver;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingPropertiesRoot;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesRoot;
 import org.opentestsystem.rdw.olap.stream.AggregateRequestProcessor;
 import org.opentestsystem.rdw.reporting.common.configuration.ReportingSystemProperties;
 import org.opentestsystem.rdw.reporting.common.configuration.ReportingSystemPropertiesResolver;
