+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/service/impl/CustomAggregateReportQueryHandler.java
@@ -1,6 +1,6 @@
 package org.opentestsystem.rdw.olap.service.impl;
 
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingProperties;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingProperties;
 import org.opentestsystem.rdw.olap.model.AggregateOrganizationQuery;
 import org.opentestsystem.rdw.olap.model.CustomAggregateReportResult;
 import org.opentestsystem.rdw.olap.model.CustomAggregateReportResult.Builder;
