+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/service/impl/DefaultUserAssessmentTypeService.java
@@ -1,6 +1,6 @@
 package org.opentestsystem.rdw.olap.service.impl;
 
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingProperties;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingProperties;
 import org.opentestsystem.rdw.reporting.common.model.CodedEntity;
 import org.opentestsystem.rdw.olap.service.AssessmentTypeService;
 import org.opentestsystem.rdw.olap.service.UserAssessmentTypeService;
