+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/service/impl/TargetReportQueryHandler.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.olap.service.impl;
 
 import org.opentestsystem.rdw.common.model.AssessmentType;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingProperties;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingProperties;
 import org.opentestsystem.rdw.olap.model.SubjectAssessment;
 import org.opentestsystem.rdw.olap.model.TargetOrganizationQuery;
 import org.opentestsystem.rdw.olap.model.TargetReportResult;
