+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/AggregateReportingPropertiesTenantIT.java
@@ -2,7 +2,7 @@ package org.opentestsystem.rdw.olap;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingProperties;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingProperties;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
