+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/AggregateReportingPropertiesTenantTest.java
@@ -3,7 +3,7 @@ package org.opentestsystem.rdw.olap;
 import com.google.common.collect.ImmutableSet;
 import org.junit.Before;
 import org.junit.Test;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingPropertiesTenant;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesTenant;
 
 import static org.assertj.core.api.Assertions.assertThat;
