+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/TestAppConfig.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.olap;
 
 
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingPropertiesTenant;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesTenant;
 import org.opentestsystem.rdw.olap.repository.DataSourceConfiguration;
 import org.opentestsystem.rdw.olap.sqlbuilder.impl.QueryProviderRepositoryHelper;
 import org.opentestsystem.rdw.reporting.common.service.impl.DefaultUserStudentFieldService;
