+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/service/impl/AbstractAggregateQueryHandlerTest.java
@@ -3,7 +3,7 @@ package org.opentestsystem.rdw.olap.service.impl;
 import com.google.common.collect.ImmutableSet;
 import org.junit.Before;
 import org.junit.Test;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingProperties;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingProperties;
 import org.opentestsystem.rdw.olap.model.AggregateReportQueryDescription;
 import org.opentestsystem.rdw.olap.model.AggregateRowReportResult;
 import org.opentestsystem.rdw.olap.model.OrganizationTypeQuery;
