+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/service/impl/ClaimReportQueryHandlerTest.java
@@ -5,8 +5,7 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingProperties;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingPropertiesTenant;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingProperties;
 import org.opentestsystem.rdw.olap.model.ClaimOrganizationQuery;
 import org.opentestsystem.rdw.olap.model.ClaimReportResult;
 import org.opentestsystem.rdw.olap.repository.ClaimReportRepository;
