+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/service/impl/CustomAggregateReportQueryHandlerTest.java
@@ -2,7 +2,7 @@ package org.opentestsystem.rdw.olap.service.impl;
 
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingProperties;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingProperties;
 import org.opentestsystem.rdw.olap.model.AggregateOrganizationQuery;
 import org.opentestsystem.rdw.olap.model.CustomAggregateReportResult;
 import org.opentestsystem.rdw.olap.repository.CustomAggregateReportRepository;
