+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/service/impl/LongitudinalReportQueryHandlerTest.java
@@ -3,7 +3,7 @@ package org.opentestsystem.rdw.olap.service.impl;
 
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingProperties;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingProperties;
 import org.opentestsystem.rdw.olap.model.LongitudinalOrganizationQuery;
 import org.opentestsystem.rdw.olap.model.LongitudinalReportResult;
 import org.opentestsystem.rdw.olap.repository.LongitudinalReportRepository;
