+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/configuration/AggregateReportingProperties.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.olap.configuration;
+package org.opentestsystem.rdw.reporting.common.configuration;
 
 import com.google.common.collect.ImmutableSet;
