+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/configuration/AggregateReportingPropertiesRoot.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.olap.configuration;
+package org.opentestsystem.rdw.reporting.common.configuration;
 
 import org.opentestsystem.rdw.multitenant.TenantConfigurationStatus;
 import org.opentestsystem.rdw.multitenant.TenantConfigurationStatusProvider;
