+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/configuration/AggregateReportingPropertiesTenant.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.olap.configuration;
+package org.opentestsystem.rdw.reporting.common.configuration;
 
 import java.util.Set;
