+++ common/src/test/java/org/opentestsystem/rdw/reporting/common/test/ItApplicationConfiguration.java
@@ -3,10 +3,13 @@ package org.opentestsystem.rdw.reporting.common.test;
 import org.opentestsystem.rdw.multitenant.TenantIdResolver;
 import org.opentestsystem.rdw.multitenant.TenantKeyResolver;
 import org.opentestsystem.rdw.multitenant.TenantProperties;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesResolver;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesRoot;
 import org.opentestsystem.rdw.reporting.common.configuration.ReportingSystemPropertiesImpl;
 import org.opentestsystem.rdw.reporting.common.multitenant.ReportingTenantIdResolver;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.cloud.context.config.annotation.RefreshScope;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -38,6 +41,18 @@ public class ItApplicationConfiguration {
     public TenantKeyResolver tenantKeyResolver(TenantProperties tenantProperties,
                                                TenantIdResolver tenantIdResolver) {
         return new TenantKeyResolver(tenantProperties, tenantIdResolver);
+    }
+
+    @Bean
+    @ConfigurationProperties("aggregate-reporting")
+    AggregateReportingPropertiesRoot aggregateReportingPropertiesRoot() {
+        return new AggregateReportingPropertiesRoot();
+    }
+
+    @Bean
+    AggregateReportingPropertiesResolver aggregateReportingProperties(TenantKeyResolver tenantKeyResolver,
+                                                                      AggregateReportingPropertiesRoot aggregateReportingPropertiesRoot) {
+        return new AggregateReportingPropertiesResolver(tenantKeyResolver, aggregateReportingPropertiesRoot);
     }
 
 }
