+++ common/src/test/java/org/opentestsystem/rdw/reporting/common/test/RepositoryIT.java
@@ -1,7 +1,6 @@
 package org.opentestsystem.rdw.reporting.common.test;
 
 import org.opentestsystem.rdw.reporting.common.jdbc.SecurityParameterProvider;
-import org.opentestsystem.rdw.reporting.common.repository.DataSourceConfiguration;
 import org.opentestsystem.rdw.utils.YamlPropertiesConfigurator;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
